Operations on molecules

Various functions to perform operations on molecules.

get.exact.mass returns the exact mass of a molecule get.natural.mass returns the natural exact mass of a molecule convert.implicit.to.explicit converts implicit hydrogens to explicit hydrogens. This function does not return any value but rather modifies the molecule object passed to it is.neutral returns TRUE if all atoms in the molecule have a formal charge of 0, otherwise FALSE

Details

In some cases, a molecule may not have any hydrogens (such as when read in from an MDL MOLfile that did not have hydrogens). In such cases, convert.implicit.to.explicit will add implicit hydrogens and then convert them to explicit ones. In addition, for such cases, make sure that the molecule has been typed beforehand.

Usage

get.exact.mass(mol) get.natural.mass(mol) convert.implicit.to.explicit(mol) is.neutral(mol)

Arguments

mol A jobjRef representing an IAtomContainer or IMolecule object

Value

get.exact.mass returns a numeric get.natural.mass returns a numeric convert.implicit.to.explicit has no return value is.neutral returns a boolean.

See also

Author

Rajarshi Guha (rajarshi.guha@gmail.com)